+++ pint2vela/examples/run_example_emcee.py
@@ -21,7 +21,7 @@ parfile, timfile = sys.argv[1], sys.argv[2]
 m, t = get_model_and_toas(parfile, timfile)
 
 # %%
-mv, tv = read_model_and_toas(parfile, timfile)
+mv, tv = read_model_and_toas(parfile, timfile, cheat_prior_scale=10)
 lnpost = vl.get_lnpost_func(mv, tv)
 prior_transform = vl.get_prior_transform_func(mv)
 
@@ -60,12 +60,13 @@ for pname, mean, std in zip(param_names, means, stds):
     print(f"{pname}\t\t{mean}\t\t{std}")
 
 # %%
-param_labels = [f"\n\n{pname}\n({m[pname].units})\n" for pname in param_names]
+# param_labels = [f"\n\n{pname}\n({m[pname].units})\n" for pname in param_names]
+param_labels = vl.get_free_param_labels(mv)
 fig = corner.corner(
     samples_v,
     labels=param_labels,
     label_kwargs={"fontsize": 11},
-    range=[0.999999] * ndim,
+    range=[0.999] * ndim,
     truths=maxlike_params_v / scale_factors,
     plot_datapoints=False,
     hist_kwargs={"density": True},